Refined the warning message and the way Character.move handles stuff.
