+++ .idea/workspace.xml
@@ -19,19 +19,12 @@
     <select />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="de0d31bb-7a3b-400a-bc5c-7b6cd08e10df" name="Default Changelist" comment="Bug fix and done some teardown.">
-      <change afterPath="$PROJECT_DIR$/src/main/java/eod/exceptions/MoveInvalidException.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/eod/specifier/condition/TargetedCondition.java" afterDir="false" />
+    <list default="true" id="de0d31bb-7a3b-400a-bc5c-7b6cd08e10df" name="Default Changelist" comment="Added Dodge card">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/eod/Character.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/Character.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/Deck.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/Deck.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/Game.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/Game.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/GameLosingException.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/exceptions/GameLosingException.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/eod/Gameboard.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/Gameboard.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/Player.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/Player.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/card/concrete/Dodge.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/card/concrete/Dodge.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/specifier/Accessing.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/specifier/Accessing.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/eod/specifier/condition/Conditions.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/specifier/condition/Conditions.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/eod/card/concrete/EmergencyHeal.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/card/concrete/action/EmergencyHeal.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/eod/card/concrete/conditional/Dodge.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/eod/card/concrete/conditional/Dodge.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -72,7 +65,8 @@
       <workItem from="1587985049498" duration="2940000" />
       <workItem from="1587990939722" duration="1192000" />
       <workItem from="1588045242119" duration="1441000" />
-      <workItem from="1588070137934" duration="4484000" />
+      <workItem from="1588070137934" duration="4926000" />
+      <workItem from="1588076006495" duration="1094000" />
     </task>
     <task id="LOCAL-00001" summary="Card disposal at the start of the game.">
       <created>1587985528937</created>
@@ -95,7 +89,14 @@
       <option name="project" value="LOCAL" />
       <updated>1587987658716</updated>
     </task>
-    <option name="localTasksCounter" value="4" />
+    <task id="LOCAL-00004" summary="Added Dodge card">
+      <created>1588075345276</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1588075345276</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -116,7 +117,9 @@
   <component name="VcsManagerConfiguration">
     <MESSAGE value="Card disposal at the start of the game." />
     <MESSAGE value="Bug fix and done some teardown." />
-    <option name="LAST_COMMIT_MESSAGE" value="Bug fix and done some teardown." />
+    <MESSAGE value="Added Dodge card" />
+    <MESSAGE value="Refined the warning message and the way Character.move handles stuff." />
+    <option name="LAST_COMMIT_MESSAGE" value="Refined the warning message and the way Character.move handles stuff." />
   </component>
   <component name="WindowStateProjectService">
     <state x="895" y="173" width="470" height="723" key="#com.intellij.ide.util.MemberChooser" timestamp="1588070268139">
@@ -131,9 +134,9 @@
       <screen x="84" y="27" width="1836" height="1053" />
     </state>
     <state x="624" y="41" width="657" height="627" key="#com.intellij.refactoring.typeMigration.ui.FailedConversionsDialog/84.27.1836.1053@84.27.1836.1053" timestamp="1587985439527" />
-    <state x="496" y="225" width="1002" height="663" key="Vcs.Push.Dialog.v2" timestamp="1587987752269">
+    <state x="496" y="225" width="1002" height="663" key="Vcs.Push.Dialog.v2" timestamp="1588075363077">
       <screen x="84" y="27" width="1836" height="1053" />
     </state>
-    <state x="496" y="225" width="1002" height="663" key="Vcs.Push.Dialog.v2/84.27.1836.1053@84.27.1836.1053" timestamp="1587987752269" />
+    <state x="496" y="225" width="1002" height="663" key="Vcs.Push.Dialog.v2/84.27.1836.1053@84.27.1836.1053" timestamp="1588075363077" />
   </component>
 </project>
