+++ src/main/java/eod/Gameboard.java
@@ -17,20 +17,24 @@ public class Gameboard implements Snapshotted, GameObject {
         board[0][0].isAttacked = true;
     }
 
-    public boolean hasObjectOn(int x, int y) throws ArrayIndexOutOfBoundsException{
+    public boolean hasObjectOn(int x, int y) throws ArrayIndexOutOfBoundsException {
         if(x<0 || x>= boardSize || y<0 || y>= boardSize) {
-            throw new ArrayIndexOutOfBoundsException("Trying to find an element at ("+x+", "+y+") on the board");
+            throw new ArrayIndexOutOfBoundsException("Trying to find an element at ("+x+", "+y+") on the board.");
         }
         return board[x][y]!=null;
     }
 
-    public void moveElement(Point from, Point to) throws MoveInvalidException {
+    public void moveElement(Point from, Point to) throws MoveInvalidException, ArrayIndexOutOfBoundsException {
         if(board[to.x][to.y] != null) {
-            throw new MoveInvalidException("There's already a character on "+to.x+", "+to.y);
+            throw new MoveInvalidException("There's already a character on ("+to.x+", "+to.y+").");
         }
-        Character temp = board[from.x][from.y];
+
+        if(to.x < 0 || to.x >= boardSize || to.y < 0 || to.y >= boardSize) {
+            throw new ArrayIndexOutOfBoundsException("Trying to move a character to ("+to.x+", "+"to.y"+").");
+        }
+
+        board[to.x][to.y] = board[from.x][from.y];
         board[from.x][from.y] = null;
-        board[to.x][to.y] = temp;
     }
 
     @Override
